Move wizard component files
