+++ client/src/components/WizardSelectButton.tsx
@@ -1,8 +1,8 @@
 import React, { FC, ReactElement, MouseEvent, useState } from 'react';
 import { Button } from 'semantic-ui-react';
 
-import CheckSVG from '../../assets/check.svg';
-import '../../css/WizardSelectButton.css';
+import CheckSVG from '../assets/check.svg';
+import '../css/WizardSelectButton.css';
 
 interface IProps {
   onClick(e: MouseEvent<HTMLButtonElement>): void;
